Switch to Svelte
