+++ .gitignore
@@ -1,3 +1,7 @@
+/public/build/
+
+.DS_Store
+
 # Dependencies
 /node_modules
