--- app/app.js
@@ -1,8 +1,0 @@
-angular.module('Player', [
-    'ngRoute',
-    'Player.player'
-  ])
-  .config(['$routeProvider', function ($routeProvider) {
-    $routeProvider
-    .otherwise({redirectTo: '/player'})
-  }])
