--- app/index.html
@@ -1,33 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <link rel="stylesheet" href="../node_modules/bootstrap/dist/css/bootstrap.min.css">
-    <script src="./assets/all.js"></script>
-
-    <!-- <script defer src="fontawesome-all.min.js"></script> -->
-    <link rel="stylesheet" href="style/style.css">
-    <title>Dusk Player</title>
-</head>
-
-<body ng-app="Player">
-    <div class="container-fluid">
-        <div ng-view></div>
-    </div>
-
-    <script src="../node_modules/angular/angular.js"></script>
-    <script src="../node_modules/angular-route/angular-route.min.js"></script>
-    <script src="../node_modules/howler/dist/howler.min.js"></script>
-    <script src="./assets/Vibrant.min.js"></script>
-
-
-    <script src="player/player.js"></script>
-    <script src="app.js"></script>
-
-
-</body>
-
-</html>
