--- app/player/player.html
@@ -1,121 +1,0 @@
-<div class="row">
-    <input class="form-control" id="search" type="search" ng-model="search.name" ng-show="playListVisible"
-        placeholder="Search" aria-label="Search">
-    <div id="playlist" ng-show="playListVisible" class="list-group style-3">
-        <!-- <input type="text" ng-model="search.name"> -->
-        <a href="" class="list-group-item list-group-item-action list" ng-repeat="song in player.playlist | filter:search"
-            ng-click="playPlaylistSong(song.index)">
-            <h6>
-                {{song.name}} • 
-                <small class="text-muted">{{song.artist}}</small>
-            </h6>
-            <!-- {{song.name}} • <em>{{song.artist}}</em> -->
-        </a>
-    </div>
-    <div class="col-5  my-auto">
-            <div class="spinner-border text-danger centerBlock" style="width: 5rem; height: 5rem;" ng-show="loading" role="status">
-                <span class="sr-only">Loading...</span>
-            </div>
-            <img id="picture" ng-show="!loading">
-    </div>
-
-    <div class="col">
-        <div class="row">
-        
-            <!-- <div id="picture"></div> -->
-            <div class="col-md-12 text-center">
-                <div id="title">
-                    <h1 id="track"> {{trackName}} </h1>
-                    <h3 id="artist"> {{trackArtist}} • 
-                        <small id="album" class="text-muted">{{trackAlbum}}</small>
-                    </h3><br>
-        
-                </div>
-            </div>
-        
-            <div class="col-md-12 text-center">
-                <div id="loading"></div>
-                <button type="button" id="prevBtn" class="btn btn-primary-outline btn-lg" ng-click="prevSong()">
-                    <i class="fas fa-step-backward fa-2x fa-inverse"></i>
-                </button>
-                <button type="button" id="playBtn" class="btn btn-primary-outline btn-lg" ng-show="!songPlaying"
-                    ng-click="playMusic()">
-                    <!-- <i class="fas fa-play fa-2x fa-inverse"></i> -->
-                    <i class="far fa-play-circle fa-4x fa-inverse"></i>
-                </button>
-                <button type="button" id="pauseBtn" class="btn btn-primary-outline btn-lg" ng-show="songPlaying"
-                    ng-click="playMusic()">
-                    <!-- <i class="fas fa-pause fa-2x fa-inverse"></i> -->
-                    <i class="far fa-pause-circle fa-4x fa-inverse"></i>
-        
-                </button>
-                <button type="button" id="nextBtn" class="btn btn-primary-outline btn-lg" ng-click="nextSong()">
-                    <i class="fas fa-step-forward fa-2x fa-inverse"></i>
-                </button>
-        
-            </div>
-        
-            <div class="col-md-12 text-center">
-                <div id="timer">0:00 </div>
-                <div id="duration">{{timer}} </div><br>
-                <!-- <div id="bar" ng-hide="songPlaying"></div> -->
-                <!-- <div id="waveform" ng-show="songPlaying" ng-click="seekToTime($event)"></div> -->
-        
-                <div class="progress" id="seek" ng-click="seekToTime($event)">
-                    <div class="progress-bar bg-danger" style="transition: width .2s cubic-bezier(0.22, 0.61, 0.36, 1);"
-                        role="progressbar" id="progress" aria-valuemin="0" aria-valuemax="100"></div>
-                </div>
-            </div>
-            <!-- <div id="progress"></div> -->
-            <br>
-            <div class="col-md-12" id="outerCtrl">
-        
-                <div class="slidecontainer justify-content-start">
-                    <i class="fas fa-volume-down"></i>
-                    <i class="fas fa-volume-up" style="float: right"></i>
-                    <input type="range" min="0" max="100" value="100" class="slider" id="myRange">
-                </div>
-        
-        
-                <button type="button" id="playlistBtn" class="btn btn-primary-outline btn-lg justify-content-end"
-                    ng-click="showPlaylist()">
-                    <!-- <i class="fas fa-list fa-lg fa-inverse"></i> -->
-                    <i class="fas fa-bars fa-lg fa-inverse"></i>
-                </button>
-        
-                <button type="button" id="shuffleBtn" ng-show="shuffle"
-                    class="btn btn-primary-outline btn-lg justify-content-end" ng-click="toggleShuffle()">
-                    <i class="fas fa-random fa-lg fa-inverse"></i>
-                </button>
-        
-                <button type="button" id="shuffleBtn" ng-show="!shuffle"
-                    class="btn btn-primary-outline btn-lg justify-content-end" ng-click="toggleShuffle()">
-                    <i class="fas fa-sync-alt fa-lg fa-inverse"></i>
-                </button>
-        
-        
-                <button type="button" id="checkboxrn" ng-show="mute" class="btn btn-primary-outline btn-lg justify-content-end"
-                    ng-click="togglecheckbox()">
-                    <i class="fas fa-volume-off fa-lg fa-inverse"></i>
-                </button>
-        
-                <button type="button" id="checkboxrn" ng-show="!mute" class="btn btn-primary-outline btn-lg justify-content-end"
-                    ng-click="togglecheckbox()">
-                    <i class="fas fa-volume-up fa-lg fa-inverse"></i>
-                </button>
-        
-        
-            </div>
-        
-            
-        
-            <!-- <div id="playlist" ng-show="playListVisible">
-                    <div id="list">
-                        <div class="list-song" ng-repeat="song in player.playlist" ng-click="playPlaylistSong($index)">{{song.name}}</div>
-                    </div>
-                </div> -->
-        </div>
-
-    </div>
-
-</div>
