+++ public/index.html
@@ -1,0 +1,24 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+	<meta charset='utf-8'>
+	<meta name='viewport' content='width=device-width,initial-scale=1'>
+
+	<link rel="stylesheet" href="../node_modules/bootstrap/dist/css/bootstrap.min.css">
+	<script src="assets/all.js"></script>
+
+	<title>Dusk Player</title>
+
+	<link rel='icon' type='image/png' href='../build/icon.png'>
+	<link rel='stylesheet' href='global.css'>
+	<link rel='stylesheet' href='build/bundle.css'>
+
+	<script src="../node_modules/howler/dist/howler.min.js"></script>
+	<script src="./assets/Vibrant.min.js"></script>
+
+	<script defer src='build/bundle.js'></script>
+</head>
+
+<body>
+</body>
+</html>
