+++ src/main.js
@@ -1,0 +1,9 @@
+import App from './App.svelte';
+
+const app = new App({
+	target: document.body,
+	props: {
+	}
+});
+
+export default app;
